Split cert_auth_utils: move test utilities to test_utils/Separate production certificate utilities from test-only certificategeneration code. Production code (TCertificateAuthValues) stays incert_auth_utils.h, test utilities move to test_utils/test_cert_auth_utils.h. Removed excessive headers in ydb_register_node_ut.cpp
