+++ ydb/core/security/certificate_check/test_utils/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    test_cert_auth_utils.cpp
+)
+
+PEERDIR(
+    contrib/libs/openssl
+)
+
+END()
