+++ ydb/core/security/ldap_auth_provider/test_utils/ya.make
@@ -6,7 +6,7 @@ SRCS(
 
 PEERDIR(
     ydb/core/protos
-    ydb/core/security/certificate_check
+    ydb/core/security/certificate_check/test_utils
 )
 
 END()
