+++ ydb/core/security/ya.make
@@ -19,6 +19,7 @@ PEERDIR(
     ydb/core/audit
     ydb/core/base
     ydb/core/protos
+    ydb/core/security/certificate_check/test_utils
     ydb/library/aclib
     ydb/library/aclib/protos
     ydb/library/login
